[android] Fix problem with multiple account notifications. fix #1584
There was a problem with notifications for multiple accounts. First of
all, notifications for other accounts have been dismissed when they
should not be (when there was only one notification for another
account) because of the incorrect check when all counters are zero.
Second, if you had notifications for another account and you navigated
inside the app, you would get repeated notifications. That happened
because we updated (sent again) the summary notificaton and it had
sound.

For the context: Android before N had no expandable notification and
summary would "consume" individual notifications. Since N user can
access individual notifications.

To solve this I tried to not touch summary when we should not but it
would still notify the user again when we actually dismiss something.

Then I tried to adjust GROUP_ALERT_BEHAVIOR so that we don't get the
sound when we update the summary but it didn't work properly before N:
there was no sound at all. Since N it is not possible to control sound
for individual notifications. So we have a hacky condition now to
use different GROUP_ALERT_BEHAVIOR by version.

Maybe it would be better to not use summary at all for older Android
versions and have one/multiple updatable notifications. We could also
group notifications by user id but that would be a bigger change. Since
we consider improving notifications to show the sender I won't change
this too much now.