Fix recurring all-day events from another timezone, fix #1809
For repeating events we remember time zone in which event was created
to have stable iterations across time zones. Unfortunately, this does
not work with all-day events which are supposed to be in the viewer's
time zone. When there's a daylight shift in creator's time zone viewer
may have day shifted by one. This commit makes calculation use viewer's
time zone for all-day events.

It also introduces some changes for extracting times of all-day events
to make them work with passed time zone. This is necessary for tests
and might also be helpful if we introduce "timezone override" in the
future.

Some functions from DateUtils had to be duplicated in CalendarUtils to
make them timezone-aware. We cannot make this change in DateUtils
because we don't depend on Luxon in the worker.