Fix padding RSA output for encryption and signing, fix #1345
The issue is invalid padding (not cryptographic padding) of encoded data
for the RSA operations. It doesn't expect that more than one leading
number is zero while in practice two or more can be.

It removes hex conversion and adds proper padding for RSA
signature. We have workaround for our own decryption but some systems
like Conscrypt in Android might complain that the data length is
invalid.