[desktop] provide option to override the default font, close #1909
adds an "override default font" option to the desktop settings viewer
which doesn't show up when there are no overrides configured for the
current platform (currently, contains SimHei's latin and chinese name on
windows, which is installed by default).

Uses a fixed list of fonts because getting a list of installed fonts
cross-platform requires a native module like font-manager or parsing of
platform-dependent command output (which may be huge)

This is not applied automatically because it replaces the default font
of the application and we can't rely on locale to detect which font
would be correct for the user.

changing the font list in main-styles.js directly would also work, but
that's not very extensible should this problem occur with other
language/locale/platform combinations. It would also require to get the
user's choice from the main thread somehow.