Whitelabel for apps and desktop (#3001)
* Whitelabel for apps, close #2804

Give body in generated HTML transparent background

This helps with background color on mobile where we set background of
the view itself. body has built-in style of being white so when we
reload the page for a split second there's white background visible.

Theme is not applied to apps early via URL parameters.

URL rewriting is removed from desktop and instead native call is used
like in other apps. This removes all the magic from desktop URL
rewriter. For logging out we use native call, like on mobile, to be able
to load the file. Before it did not actually work and only loaded page
 after we've been failing to load index.html/login.

Now the user has a choice between built-ins and custom themes in apps.
Theme picker on login screen is now also a dropdown.

Reload function for apps is changed to use map of query parameters
instead of passing pre-built query string. This simplifies handling of
query params when we want to append a theme or noAutoLogin.

* Don't give bare ipcRenderer to the web context

* Fix parsing 3 digit colors, fix #3149

Co-authored-by: ivk <ivk@tutao.de>