[desktop] handle relative settings URLs in tutanota team mails
*  the mail's `/settings/invoice` href is interpreted as an
absolute path, the root of which is `https://mail.tutanota.com/`
in browsers (correct) and the file system root in
electron (wrong).
* changing it to a relative path, e.g. `../../settings/invoice`
would _theoretically_ work, but relies on the link being displayed
in the correct location relative to the settings page.
Practically, any link navigation in the desktop clients that's
pointing to the file system is prevented by us, because we don't
want the shell to open/execute random files from links that got sent by mail.
* the `file:///C:/settings/invoice` link is generated from the mail's
`/settings/invoice` href by electron before dispatching the onclick event.

this commit allows for relative navigation from Tutanota Team
mails if the href contains a drive letter between origin and
the settings path.

fix #3333