[android] Update Android compileSdk and libraries, close #3470
Some code cleanups are done to fix the warnings.

jcenter is removed as it is deprecated.

Activity result/permission warnings are suppressed because we can't
access requestCode with new API and we need it to identify the
request.

Some PendingIntent warnings are left as-is because FLAG_MUTABLE is not
available on older SDKs and there doesn't seem to be compat
replacement.