Remove isVisible from ButtonAttrs/NavButtonAttrs, fix #3774
isVisible() had some ad-hoc usage in a few places but for the most part
it is a leftover from old-style class components. It was misleading as
Button itself didn't use it (why render a button but then tell it not
to render?) and you needed to know where to apply it and in those
places it's better to be more explicit and *not* return a button at all.

Caching attributes is antipattern in any case and we shouldn't build
capabilities for doing it.

Also removed NavButtonAttrs as one of the options for DropdownN as it
was never used this way (it also doesn't make sense conceptually).