Change date parsing to not remove empty date parts, fix #3782
DateParser._cleanUpAndSplit removes any characters
that cannot be dealt with, such as the char 'a'.
This results in an empty date part if a date of the form
"a/4/12", for example, is entered by a user.
_cleanUpAndSplit would then remove the part containing only 'a'
(zero-length) and the date would be interpreted as
month/day or day/month, depending on locale, which
resulted in unexpected dates being saved.

Instead, we cut off dates that have more than three
parts and do not remove zero-length parts, which
fixes the test failure mentioned in the issue.