Fix search date picker on mobile, fix #3860
Date selection for the search index range is presented in a dialog. By
default, our dialogs try to move focus to the most relevant element.
Element with the highest priority is a text field.

This is not something that we want on mobile because on mobile we only
want to display the date picker field and on interaction we actually
want to show native date picker. That's why we render invisible
input[type=date] over our textField.

When dialog was focusing the field though we were showing non-mobile
date picker which we didn't account for. With this change we prevent
automatic focus of date field and therefore no desktop date picker will
be shown.