Reconnect webSocket when sleep is detected, fix #3449
It seems like the issue is caused by how Chromium handles suspension
(or rather doesn't). After coming back from suspension of the device it
doesn't detect that the connection might have been broken until ping
interval expires (which is minutes for us). In this time websocket is in
the peculiar state where it thinks that it is connected, but it missed
some events that were sent by the server but not acknowledged by the
client. Once ping interval occurs the server would re-send all the
events missed by the client which is what we would want ideally, but it
seems to be too slow to be reliable. In this time users don't see the
new state and the results of their actions are not displayed. This is
a suboptimal experience.

To address this we introduced SleepDetector. It uses simple idea with
scheduling a recurring task and measuring the time between invocations.
Our time period is hopefully loose enough that slight throttling by
browsers won't affect it. Once we realize that more time than needed
has passed we force reconnect.

Alternatively, we could try to send an event over the socket to force
re-synchronization, but it might not produce the result we want,
restarting the connection seems to be the most reliable method.