Only store desired data in offline mode, #3992
We now clear any old or unwanted data from the offline database at login
 - all mails and mailbodies from the trash and spam folders
 - any mails and mailbodies older than a certain cutoff date

The cutoff date is configurable in email settings, and stored in the
offline database metadata table itself.

We still write to the offline database and only clear it on login. We
might want to improve this somewhat - cleanup periodically or after big
writes? if so we probably don't want to compact the database every time.