Make sure contacts are always available in offline mode, fix #4134
When the indexer is initialized on login it previously only downloaded
contacts if they have never been indexed before. Now we additionally
download (and therefore cache) them if we have a new offline db.

Note: When logging in offline the contact list will now always be
available but suggestions still don't work as we can only initialize
index db once we have a usergroupkey.