[desktop] ignore closed db error when in expired session recovery mode
when the saved credentials have a session that expired (ie deleted from
other device) we detect this and reset the session. this includes
closing the websocket asynchronously and closing the offline DB.

the app may try to write to the db after closing it and before closing
the websocket, which normally would not be allowed, but is fine to
ignore in this state since we're going to recreate the session anyway.

close #4401

Co-authored-by: ivk <ivk@tutao.de>