[android] add some values to the MimeTypeMap
On old androids, webkit's MimeTypeMap would return null for
some extensions used in our assets (ie .js).

since we proxy the requests for assets, we set the mimetype
manually with mimetypemap, which doesn't sit well with the
WebView if the mimetype is null (refuses to load it as executable).

Because of that worker would never be initialized (it loads polyfill.js
dynamically). For users it would look like the request is stuck.

close #4453

Co-authored-by: ivk <ivk@tutao.de>