Suppress paste events and prevent editing if the editor is disabled.
There was an issue where the signature preview could be edited unless
it is disabled later by changing the preview, so this change disables
it when it is initialized if needed.

Squire also handled pasting before our code could handle it. Luckily,
Squire has a willPaste event which can be used to suppress paste
events, so this change utilizes this and removes the "not working
currently" comment.

Fixes #2690