[android] fix crash due to devtools re-requesting login page
when devtools is opened, the webview re-requests the login page with
an url like `assets.tutanota.com/index-app.html/login`, which didn't
resolve to any MimeType and crashed the app.

We're now explicitly handling URLs that start with the index-app.html
and a slash and give them text/html MimeType.