Fix not being able to open emails if the list is cannot be loaded
Easiest to reproduce on mobile: when only part of the list is loaded
successfully try to select an email. It won't do anything because it
is waiting for the list loading before opening the email.

We could not find any reason on why it was done, it was there from the
very beginning. It is already in the list if it's selected, the only
possible problem is animation perf.

We also removed try/catch around it as there shouldn't be anything
fallible in there anymore.