Reset cached DOM elements in List on remove
fix #4515

The issues occurred because the list was taken out of the tree and
rendered as a completely new element but the deferred DOM elements
for the loading indicator wouldn't get new values because promise can
only be resolved once (it is in fact an error to try to do it multiple
times).

Now that indicator is reset on remove we will use the new indicator and
it will be updated appropriately.