Fix double login in specific cases
Double login could occur if LoginView would start the normal login
process and then the URL would change (e.g. when native part would
change trigger login with specific account). In this case a second
LoginView would be constructed and would start the login on its own.
Normally this shouldn't happen but for LoginView we specifically
opted into not caching it to not accidentally hold to credentials any
longer than necessary.

To prevent this issue we should not create LoginView manually but let
mithril do it. This is a bit tricky as LoginView needs some info passed
into it. We ended up rewriting LoginView as a proper component and
creating a new RouteResolver implementation that can work with this
type of views.

fix #4554