Fix unscheduling alarms in web version
CalendarModel receives a function to get alarm scheduler async and
lazily. It doesn't cache it internally and would request a new one
each time. When it would try to unschedule the alarms it wouldn't find
anything about existing timeout ids.

We make sure to cache the scheduler in MainLocator so there will be one
instance only.

fix #4676