Guard against missing translation keys in the feature list
the feature list is maintained on the website repository, which means
that new features are added there.
this can lead to a situation where the new translation key is in the
project, but currently deployed clients don't have it in the lang files.

we don't show the unknown feature list items in this case.

#4670