fix reminder not rendered while editing existing calendar event
especially noticeable on android, but also appears on other devices
sometimes.
The CalendarEventViewModel loads exiting calendar data asynchronously,
this also applies to reminders. The CalendarEventViewModel constructor
now gets a callback that is called after the asynchronous loading is
finished. Passing m.redraw() as callback updates the UI and displays the
events.

fix #4544