Add "autocomplete" to login input fields to better support autofill
androids example autofill app needs the "autocomplete" field in an
input field to offer autofill. The value of the autocomplete field is an
instruction how this field should be handled.

There are other possible "autocomplete" values like "new-password"
which can be used as well. Wherever TextField is created, the
"autocomplete" field will now be filled with the appropriate value,
if reasonable.