Fix calendar layout bugs, fix #4847
Changes made in 8bc7cf27 did not take into account the cases where
events take up the whole column, so it would check for visual overlaps
in all day & month cases where it's not necessary, and it would try to
introduce more columns when events would be next to each other in one
column (like same row but next day in month view). Now we only do it
conditionally.

Co-authored-by: ivk <ivk@tutao.de>