fix show progress for download all in browser
previously only downloading a single file would show a blocking progress
indication. Now this progress indication is also shown for the save all
operation. If the download of a single file fails, an error message is
displayed, mentioning this specific file. Nevertheless, the downloading
process for the other files will continue in the background.
Furthermore, the progress indication also shows a completeness
indication in form of a bar.

#4811