[desktop] make offline db creation more robust
we're getting reports of clients trying to open databases while another
database is already opened on the window, which should not happen.

this change to getOrCreateDb immediately assigns the promise of the new
database to the cache instead of waiting for sqlcipher to finish
initializing, which should make the method more robust against
concurrent calls.

also adds userIds to the error message so we can see if it's the same
users database or different users in case it goes wrong anyway.

#4866