Fix offline DB being deleted every time
DB downgrade check was not implemented properly where it would
determine runtime version based on migrations. We run into a situation
where we had multiple model versions with no migrations which made the
check think that we are downgrading every time. Switching it to using
modelInfo fixed the issue.

Another issue was not running migrations (and non populating metadata)
after DB re-creation. We've added the code to run the migrator on the
re-created database and handling for the old invalid state.

fix #5044