Fix incorrectly copying event when adding exclusions
The issue happened because we had a new code path for creating event:
when adding an exclusion we copy most of the event. However, clone()
also copies some hidden technical fields. When we encrypt the instance
we will pick them up and use them instead of encrypting the whole
instance again. We cannot do that when we have a new session key.

To fix this we now remove all the technical fields from the instance
before encrypting it.

fix #5317